pyproject.toml
[project]
name = "strands-agents-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.38.23",
    "psycopg>=3.2.9",
    "strands-agents>=0.1.4",
    "strands-agents-tools>=0.1.3",
]
